chore: delete tauri-bootstrap.ts + tauri-init.ts, create clean bootstrap.ts
- Replaced tauri-bootstrap.ts (285 lines, half dead Tauri code) with
  bootstrap.ts (140 lines, CEF-only)
- Deleted tauri-init.ts (dead — only imported by old bootstrap)
- Updated index.html entry point: tauri-bootstrap.ts → bootstrap.ts
- Removed @tauri-apps/cli devDependency (Tauri build tool, unused)
- Cleaned up wave.ts comments referencing tauri-bootstrap
- Added cleanup audit report
