+++ index.html
@@ -49,7 +49,7 @@
     <link rel="stylesheet" href="/fontawesome/css/sharp-solid.min.css" />
     <link rel="stylesheet" href="/fontawesome/css/sharp-regular.min.css" />
     <link rel="stylesheet" href="/fontawesome/css/custom-icons.min.css" />
-    <script type="module" src="frontend/tauri-bootstrap.ts"></script>
+    <script type="module" src="frontend/bootstrap.ts"></script>
   </head>
   <body class="init" data-colorscheme="dark">
     <div id="startup-loading"><svg xmlns="http://www.w3.org/2000/svg" aria-hidden="true" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 1200 1200">
