+++ package.json
@@ -30,7 +30,6 @@
     "@eslint/js": "^8.57.0",
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@tailwindcss/vite": "^4.1.18",
-    "@tauri-apps/cli": "2.10.0",
     "@types/color": "^4.2.0",
     "@types/css-tree": "^2",
     "@types/debug": "^4",
